+++ source/Support/Stat.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 
 namespace ProveAA.Support {
 	class Stat {
-		public byte Base;
-		public byte Current;
+		public short Base;
+		public short Current;
 
 		public override string ToString() {
 			return $"{Current} ({Base})";
