+++ source/Trap/HealthTrap.cs
@@ -8,7 +8,7 @@ namespace ProveAA.Trap {
 	class HealthTrap : BasicTrap {
 		public override bool PlayerStepIn(Creature.Player player) {
 			PrintText("HP trap!");
-			player.hitPoints.Current -= (byte)(player.hitPoints.Current / 4);
+			player.hitPoints.Current -= player.hitPoints.Current / 4;
 			return true;
 		}
 	}
